refactoring
move ISiteContextResolver back into models project, change PathString to
string to avoid http dependency
