+++ src/cloudscribe.Core.Models/cloudscribe.Core.Models.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <Description>model classes for cloudscribe core. you should not normally install this separately, it will be installed if you install cloudscribe.Core.Web</Description>
-    <VersionPrefix>2.0.17</VersionPrefix>
-    <AssemblyVersion>2.0.17</AssemblyVersion>
+    <VersionPrefix>2.0.18</VersionPrefix>
+    <AssemblyVersion>2.0.18</AssemblyVersion>
     <Authors>Joe Audette</Authors>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
     <AssemblyName>cloudscribe.Core.Models</AssemblyName>
