+++ src/cloudscribe.Core.Web.Mvc/cloudscribe.Core.Web.Mvc.csproj
@@ -2,8 +2,8 @@
 
   <PropertyGroup>
     <Description>MVC Controllers for cloudscribe.Core.Web</Description>
-    <VersionPrefix>2.1.22</VersionPrefix>
-    <AssemblyVersion>2.1.22</AssemblyVersion>
+    <VersionPrefix>2.1.23</VersionPrefix>
+    <AssemblyVersion>2.1.23</AssemblyVersion>
     <Authors>Joe Audette</Authors>
     <PackageIconUrl>https://raw.githubusercontent.com/joeaudette/cloudscribe/master/cloudscribe-icon-32.png</PackageIconUrl>
     <PackageProjectUrl>https://github.com/joeaudette/cloudscribe</PackageProjectUrl>
