+++ src/cloudscribe.Core.Web/Components/SiteResolver.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Author:              Joe Audette
 // Created:             2016-02-04
-// Last Modified:       2018-03-13
+// Last Modified:       2018-04-26
 // 
 
 using cloudscribe.Core.Models;
