+++ src/sourceDev.WebApp/appsettings.json
@@ -94,7 +94,7 @@
     "RelatedSiteId": "00000000-0000-0000-0000-000000000000"
   },
   "SiteConfigOptions": {
-    "DefaultTheme": "",
+    "DefaultTheme": "bootstrap4",
     "FirstSiteTheme": "",
     "RolesThatCannotBeDeleted": ""
   },
